+++ spec/test/coverage_spec.rb
@@ -5,12 +5,13 @@ RSpec.describe OpenapiFirst::Test::Coverage do
   let(:definition) { OpenapiFirst.load(filepath) }
 
   before(:each) do
+    described_class.install
     OpenapiFirst::Test.register(filepath)
     described_class.start
   end
 
   after(:each) do
-    described_class.stop
+    described_class.uninstall
     described_class.reset
   end
 
@@ -22,17 +23,21 @@ RSpec.describe OpenapiFirst::Test::Coverage do
 
   let(:result) { described_class.result }
 
-  describe '.start' do
-    after { described_class.stop }
-
+  describe '.install' do
     it 'installs global hooks' do
+      described_class.install
+
       hooks = OpenapiFirst.configuration.hooks
-      described_class.stop
-      expect(hooks[:after_request_validation]).to be_empty
-      expect(hooks[:after_response_validation]).to be_empty
-      described_class.start
       expect(hooks[:after_request_validation]).not_to be_empty
       expect(hooks[:after_response_validation]).not_to be_empty
+    end
+
+    it 'does not install hooks multiple times' do
+      2.times { described_class.install }
+
+      hooks = OpenapiFirst.configuration.hooks
+      expect(hooks[:after_request_validation].count).to eq(1)
+      expect(hooks[:after_response_validation].count).to eq(1)
     end
   end